Add example forwarding stripped messages
